replace cjs with mjs scripts
